+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
---
+- Add `ParabolaFit` feature: curvature, extremum value, and reduced $\chi^2$ of a
+  weighted parabolic (quadratic) least-squares fit
+  https://github.com/light-curve/light-curve-feature/issues/166
 
 ### Changed